HistoryToken.setAnchoredSpreadsheetSelection clears action when different
