+++ src/main/java/walkingkooka/spreadsheet/dominokit/history/SpreadsheetRowInsertBeforeHistoryToken.java
@@ -98,8 +98,7 @@ public class SpreadsheetRowInsertBeforeHistoryToken extends SpreadsheetRowInsert
                         count
                 );
         context.pushHistoryToken(
-                previous.clearAction()
-                        .setAnchoredSelection(
+                previous.setAnchoredSelection(
                                 previous.anchoredSelectionOrEmpty()
                                         .map(
                                                 a -> a.selection()
