+++ src/main/java/walkingkooka/spreadsheet/dominokit/ui/viewport/SpreadsheetViewportComponentTableCellHeader.java
@@ -47,7 +47,6 @@ abstract class SpreadsheetViewportComponentTableCellHeader<T extends Spreadsheet
         super();
 
         final HistoryTokenAnchorComponent anchor = context.historyToken()
-                .clearAction()
                 .setAnchoredSelection(
                         Optional.of(
                                 selection.setDefaultAnchor()
