+++ src/test/java/walkingkooka/spreadsheet/dominokit/history/SpreadsheetCellHighlightHistoryTokenTestCase.java
@@ -18,8 +18,6 @@
 package walkingkooka.spreadsheet.dominokit.history;
 
 import org.junit.jupiter.api.Test;
-import walkingkooka.spreadsheet.reference.AnchoredSpreadsheetSelection;
-import walkingkooka.spreadsheet.reference.SpreadsheetSelection;
 
 public abstract class SpreadsheetCellHighlightHistoryTokenTestCase<T extends SpreadsheetCellHighlightHistoryToken> extends SpreadsheetCellHistoryTokenTestCase<T> {
 
@@ -30,39 +28,6 @@ public abstract class SpreadsheetCellHighlightHistoryTokenTestCase<T extends Spr
     @Test
     public void testSetMenuWithCell() {
         this.setMenuWithCellAndCheck();
-    }
-
-
-    @Test
-    public final void testSetAnchoredSelectionDifferentColumn() {
-        final AnchoredSpreadsheetSelection different = SpreadsheetSelection.parseColumn("B")
-                .setDefaultAnchor();
-
-        this.setAnchoredSelectionAndCheck(
-                this.createHistoryToken(),
-                different,
-                HistoryToken.column(
-                        ID,
-                        NAME,
-                        different
-                )
-        );
-    }
-
-    @Test
-    public final void testSetAnchoredSelectionDifferentRow() {
-        final AnchoredSpreadsheetSelection different = SpreadsheetSelection.parseRow("2")
-                .setDefaultAnchor();
-
-        this.setAnchoredSelectionAndCheck(
-                this.createHistoryToken(),
-                different,
-                HistoryToken.row(
-                        ID,
-                        NAME,
-                        different
-                )
-        );
     }
 
     // patternKind......................................................................................................
